src/requirements.txt
slackclient
python-dotenv
flask
slackeventsapi
pylint